+++ dandy/llm/service/service.py
@@ -25,10 +25,14 @@ class Service:
     def __init__(
             self,
             config: BaseLlmConfig,
+            context_length: Union[int, None] = None,
+            max_completion_tokens: Union[int, None] = None,
             seed: Union[int, None] = None,
             temperature: Union[float, None] = None):
 
         self._config = config
+        self._context_length = context_length
+        self._max_completion_tokens = max_completion_tokens
         self._seed = seed
         self._temperature = temperature
 
@@ -76,8 +80,10 @@ class Service:
 
     def get_request_body(self) -> BaseRequestBody:
         return self._config.generate_request_body(
+            context_length=self._context_length,
+            max_completion_tokens=self._max_completion_tokens,
+            seed=self._seed,
             temperature=self._temperature,
-            seed=self._seed
         )
 
     def process_prompt_to_model_object(